更新查询一周内容

+++ src/main/java/scms/domain/ServerJson/EventItem.java
@@ -9,14 +9,14 @@ package scms.domain.ServerJson;
 //json 返回值
 
 public class EventItem {
-    public EventItem(int id, String title, int location, long begin, long length){
+    public EventItem(long id, String title, int location, long begin, long length){
         this.id = id;
         this.title = title;
         this.begin = begin;
         this.length = length;
         this.location = location;
     }
-    public int id;
+    public long id;
     public String title;
     public int location;
     public long begin;
